Win Condition Method Update
Added a Wins Needed method to determine the number of wins needed for a player to win.

Added a win condition check that will go through the new player list class to see if there is a winner yet.

Added a player list class to easier keep track of each game player through the app.